resources updated

+++ src/main/java/de/uniks/pioneers/controller/IngameScreenController.java
@@ -8,6 +8,7 @@ import de.uniks.pioneers.model.*;
 import de.uniks.pioneers.services.*;
 import de.uniks.pioneers.ws.EventListener;
 import io.reactivex.rxjava3.disposables.CompositeDisposable;
+import javafx.application.Platform;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.ListChangeListener;
 import javafx.collections.MapChangeListener;
@@ -18,21 +19,16 @@ import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.control.*;
 import javafx.scene.image.Image;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
 import javafx.scene.image.ImageView;
-import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
+import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.paint.ImagePattern;
+import javafx.scene.paint.Paint;
 import javafx.scene.shape.Circle;
 import javafx.scene.shape.Polygon;
-import javafx.scene.layout.VBox;
 import javafx.scene.shape.SVGPath;
-import javafx.scene.paint.Paint;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
@@ -49,8 +45,7 @@ import static de.uniks.pioneers.GameConstants.*;
 
 @Singleton
 public class IngameScreenController implements Controller {
-    private final GameService gameService;
-    private final Provider<LobbyScreenController> lobbyScreenControllerProvider;
+    @FXML public Pane root;
     @FXML public Pane turnPane;
     @FXML public SVGPath streetSVG;
     @FXML public SVGPath houseSVG;
@@ -83,10 +78,13 @@ public class IngameScreenController implements Controller {
     private final App app;
     private final Provider<RulesScreenController> rulesScreenControllerProvider;
     private final Provider<SettingsScreenController> settingsScreenControllerProvider;
+    private final Provider<LobbyScreenController> lobbyScreenControllerProvider;
     private final IngameService ingameService;
-    private final ArrayList<HexTileController> tileControllers = new ArrayList<>();
+    public final ArrayList<HexTileController> tileControllers = new ArrayList<>();
 
+    private final GameStorage gameStorage;
     private final UserService userService;
+    private final GameService gameService;
     private final EventListener eventListener;
     private final DiceSubcontroller diceSubcontroller;
 
@@ -96,14 +94,15 @@ public class IngameScreenController implements Controller {
     private final ArrayList<StreetPointController> streetPointControllers = new ArrayList<>();
     private final CompositeDisposable disposable = new CompositeDisposable();
 
-    private final GameStorage gameStorage;
-
     @Inject
     Provider<GameChatController> gameChatControllerProvider;
     @Inject
     Provider<StreetPointController> streetPointControllerProvider;
     @Inject
     Provider<IngamePlayerListElementController> elementProvider;
+    @Inject
+    Provider<IngamePlayerResourcesController> resourcesControllerProvider;
+    
     private boolean darkMode = false;
 
 
@@ -183,13 +182,20 @@ public class IngameScreenController implements Controller {
                 })
         );
 
+        Platform.runLater(() -> {
+            // init controller for player resources box
+            IngamePlayerResourcesController ingamePlayerResourcesController = resourcesControllerProvider.get();
+            ingamePlayerResourcesController.root = this.root;
+            ingamePlayerResourcesController.render();
+            ingamePlayerResourcesController.init();
+        });
+
         // add change listeners
         // players change listener
         gameService.players.addListener((MapChangeListener<? super String, ? super Player>) c -> {
-            if (c.wasAdded()) {
-                System.out.println("Player was added!");
+            if (c.wasAdded() && !c.wasRemoved()) {
                 this.renderPlayer(c.getValueAdded());
-            } else if (c.wasRemoved()) {
+            } else if (c.wasRemoved() && !c.wasAdded()) {
                 this.deletePlayer(c.getValueRemoved());
             }
         });
@@ -237,7 +243,6 @@ public class IngameScreenController implements Controller {
         ExpectedMove move = currentState.expectedMoves().get(0);
 
         this.setSituationLabel(move);
-
         if (move.players().get(0).equals(userService.getCurrentUser()._id())) {
             // enable posting move
             System.out.println("It's your turn now!");
@@ -442,8 +447,8 @@ public class IngameScreenController implements Controller {
             tile.findCorners(this.buildingControllers);
             tile.link();
         }
-        for (BuildingPointController buildingPoint : this.buildingControllers) {
 
+        for (BuildingPointController buildingPoint : this.buildingControllers) {
             // put buildingPointControllers in Hashmap to access with coordinates
             this.buildingPointControllerHashMap.put(
                     buildingPoint.generateKeyString(),
